feat: accessToken, dogId 연결

+++ app/src/main/java/com/project/meongcare/supplement/model/entities/SupplementData.kt
@@ -37,7 +37,7 @@ data class DetailSupplement(
 )
 
 data class SupplementDto(
-    val dogId: Int,
+    val dogId: Long,
     val brand: String,
     val name: String,
     val intakeCycle: Int,
